Update to cleanup and setup sripts to clean/creation of illness_docs dir only if needed

+++ cleanup.py
@@ -4,13 +4,14 @@ import os
 
 def main():
     illness_dir = os.path.join(os.path.dirname(__file__), 'illness_docs')
-    for json_file in os.listdir(illness_dir):
-        file_path = os.path.join(illness_dir, json_file)
-        try:
-            if os.path.isfile(file_path):
-                os.unlink(file_path)
-        except Exception, e:
-            print e
+    if os.path.exists(illness_dir):
+        for json_file in os.listdir(illness_dir):
+            file_path = os.path.join(illness_dir, json_file)
+            try:
+                if os.path.isfile(file_path):
+                    os.unlink(file_path)
+            except Exception, e:
+                print e
 
     r = redis.StrictRedis(host='localhost', port=6379, db=0)
     r.flushall()
